requirements.txt
Django
djangorestframework
django-filter
markdown
django-cors-headers
django-versatileimagefield
django-model-utils
djoser
requests
Pillow
drf-writable-nested
django-constance[database]
psycopg2-binary
flake8
drf-spectacular